change version from int to long

+++ server/src/main/java/org/apache/iotdb/db/wal/WALManager.java
@@ -77,7 +77,7 @@ public class WALManager implements IService {
   }
 
   public void registerWALNode(
-      String applicantUniqueId, String logDirectory, int startFileVersion, long startSearchIndex) {
+      String applicantUniqueId, String logDirectory, long startFileVersion, long startSearchIndex) {
     String s = config.getDataRegionConsensusProtocolClass();
     if (config.getWalMode() == WALMode.DISABLE
         || !config
