+++ server/src/test/java/org/apache/iotdb/db/wal/recover/WALRecoverManagerTest.java
@@ -167,7 +167,7 @@ public class WALRecoverManagerTest {
     int threadsNum = 5;
     ExecutorService executorService = Executors.newFixedThreadPool(threadsNum);
     List<Future<Void>> futures = new ArrayList<>();
-    int firstWALVersionId = walBuffer.getCurrentWALFileVersion();
+    long firstWALVersionId = walBuffer.getCurrentWALFileVersion();
     for (int i = 0; i < threadsNum; ++i) {
       IMemTable fakeMemTable = new PrimitiveMemTable();
       int memTableId = fakeMemTable.getMemTableId();
@@ -201,7 +201,7 @@ public class WALRecoverManagerTest {
     }
     Thread.sleep(1_000);
     // write normal .wal files
-    int firstValidVersionId = walBuffer.getCurrentWALFileVersion();
+    long firstValidVersionId = walBuffer.getCurrentWALFileVersion();
     IMemTable targetMemTable = new PrimitiveMemTable();
     WALEntry walEntry =
         new WALEntry(targetMemTable.getMemTableId(), getInsertRowPlan(DEVICE2_NAME, 4L), true);
@@ -225,7 +225,7 @@ public class WALRecoverManagerTest {
     int threadsNum = 5;
     ExecutorService executorService = Executors.newFixedThreadPool(threadsNum);
     List<Future<Void>> futures = new ArrayList<>();
-    int firstWALVersionId = walBuffer.getCurrentWALFileVersion();
+    long firstWALVersionId = walBuffer.getCurrentWALFileVersion();
     for (int i = 0; i < threadsNum; ++i) {
       IMemTable fakeMemTable = new PrimitiveMemTable();
       int memTableId = fakeMemTable.getMemTableId();
@@ -259,7 +259,7 @@ public class WALRecoverManagerTest {
     }
     Thread.sleep(1_000);
     // write normal .wal files
-    int firstValidVersionId = walBuffer.getCurrentWALFileVersion();
+    long firstValidVersionId = walBuffer.getCurrentWALFileVersion();
     IMemTable targetMemTable = new PrimitiveMemTable();
     InsertRowPlan insertRowPlan = getInsertRowPlan(DEVICE2_NAME, 4L);
     targetMemTable.insert(insertRowPlan);
